Also split out migration tests
